Add Hackable trait to all terminals / IFF locks / lockers
