+++ common/src/main/scala/net/psforever/objects/serverobject/mblocker/Locker.scala
@@ -3,9 +3,10 @@ package net.psforever.objects.serverobject.mblocker
 
 import akka.actor.{ActorContext, Props}
 import net.psforever.objects.GlobalDefinitions
+import net.psforever.objects.serverobject.hackable.Hackable
 import net.psforever.objects.serverobject.structures.Amenity
 
-class Locker extends Amenity {
+class Locker extends Amenity with Hackable {
   def Definition : LockerDefinition = GlobalDefinitions.mb_locker
 }
